Update dashboard + routing again
Go back to using history.push() in the login page; do redirection to "/login" if the user is not logged in within Dashboard, MealPlan, Recommendations, and Detail's useEffect()'s instead. <Redirect to="/dashboard" /> was making the login page do too many nested updates when logging in on a browser that hadn't already cached the site.