Update getData() so it reads from the Firestore
Recipe and user data is fetched from the Firestore when
the Dashboard page loads. This data is stored using Redux so
the Meal Plan and Recommendations pages don't have to fetch
it from the Firestore as well.

Other updates:
- No longer storing auth.currentUser in the AuthContext;
previously this entire object could be viewed using React's browser
dev tools, which seems dangerous because it has things like
what the user's id is in our database. Instead, we're now
importing auth so we can do auth.currentUser in components where
data about currentUser is needed.

- Remove console logs of errors for security (to avoid giving more
information than needed).

- Remove unused headerStore and PrivateRoute (which was also using
the currentUser data that was stored in AuthContext).

- Update index.html to use Savor logos.